address comments and ut fix
