+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicLogicalOperators.scala
@@ -583,16 +583,6 @@ case class Aggregate(
   }
 }
 
-case class AggregateWithHaving(
-    havingCondition: Expression,
-    child: LogicalPlan)
-  extends UnaryNode {
-
-  override lazy val resolved: Boolean = false
-
-  override def output: Seq[Attribute] = child.output
-}
-
 case class Window(
     windowExpressions: Seq[NamedExpression],
     partitionSpec: Seq[Expression],
